Scrolling hotfix

+++ src/components/admin/orders/orders.js
@@ -8,16 +8,24 @@ export default ({
     controller
 });
 
-controller.$inject = ['orderService', '$state', 'orderPickupService', 'dateService'];
+controller.$inject = ['orderService', '$state', 'orderPickupService', 'dateService', 'pickupService', '$window'];
 
-function controller(orderService, $state, orderPickupService, dateService) {
+function controller(orderService, $state, orderPickupService, dateService, pickupService, $window) {
     this.days = ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'];
     this.months = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'];
 
-    this.$onInit = () => {
+    this.$onInit = () => {  
         if(this.token) {
             this.getOrders();
             this.getPickups();
+
+            pickupService.getAll()
+                .then(data => {
+                    this.pickupLocations = data;
+                    const all = {_id: 'all', name: 'all'};
+                    this.pickupLocations.push(all);
+                    this.filterLocation = all;
+                });
         }
         else $state.go('admin.login');
         this.type = 'pickup';
@@ -30,6 +38,14 @@ function controller(orderService, $state, orderPickupService, dateService) {
             });
     };
 
+    this.logChange = () => {
+        console.log(this.filterLocation);
+    };
+
+    this.set = type => {
+        this.type = type;
+    };
+
     this.getPickups = () => {
         orderPickupService.getAll(this.token)
             .then(data => {
@@ -37,51 +53,57 @@ function controller(orderService, $state, orderPickupService, dateService) {
                     return new Date(curr.pickupDate) - new Date(next.pickupDate);
                 });
                 data.forEach(order => {
+                    order.unsavedCompleted = order.completed;
                     order.date = new Date(order.date).toDateString();
                     order.pickupDate = dateService.dateStringToObj(new Date(order.pickupDate).toDateString());
                     console.log(order.pickupDate);
                 });
                 this.pickups = data;
-                console.log(this.pickups);
+                console.log('pickups', this.pickups);
             });
     };
 
     this.removePickup = order => {
-        const index = this.orders.indexOf(order);
-        orderPickupService.deleteOrder(order._id, this.token)
-            .then(data => {
-                console.log(data);
-                this.pickups.splice(index, 1);
-            });
+        const prompt = $window.prompt('Are you sure you want to delete this order? Once it is deleted, it is gone forever! (y/n)');
+        if(prompt === 'y') {
+            const index = this.orders.indexOf(order);
+            orderPickupService.deleteOrder(order._id, this.token)
+                .then(() => {
+                    this.pickups.splice(index, 1);
+                });
+        } else {
+            return;
+        }
     };
 
     this.updatePickup = order => {
         const copy = {};
+        order.completed = order.unsavedCompleted;
+
         Object.keys(order).forEach(key => {
             copy[key] = order[key];
         });
+
+        delete copy.unsavedCompleted;
         copy.date = new Date(copy.date);
         copy.pickup = copy.pickup._id;
         copy.pickupDate = dateService.dateObjToString(copy.pickupDate);
-        console.log('copy of the order', copy);
         orderPickupService.updateOrder(copy._id, copy, this.token)
             .then(updated => console.log(updated));
     };
 
     this.removeOrder = order => {
+        $window.alert('Are you sure you want to delete this order?');
         const index = this.orders.indexOf(order);
         orderService.deleteOrder(order._id, this.token)
-            .then(data => {
-                console.log(data);
+            .then(() => {
                 this.orders.splice(index, 1);
             });
     };
     this.updateOrder = update => {
         this.calcNewTotal(update);
-        console.log('update function called', update);
         orderService.updateOrder(update._id, update, this.token)
-            .then(data => {
-                console.log(data);
+            .then(() => {
                 this.setOrderToUpdate(null);
             });
     };
@@ -100,6 +122,5 @@ function controller(orderService, $state, orderPickupService, dateService) {
     this.addItem = newItem => {
         this.orderToUpdate.items.push(newItem);
         newItem = {};
-        console.log(this.orderToUpdate.items);
     };
 }
